+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added new method ```Result::removeByValue()```.
+- Added new query condition ```Conditions\NotInArray```.
 
 ## [1.0.3] - 2023-09-23